Move hir_ty incremental test to its own file
