+++ crates/hir_ty/src/tests.rs
@@ -7,6 +7,7 @@ mod traits;
 mod method_resolution;
 mod macros;
 mod display_source_code;
+mod incremental;
 
 use std::{env, sync::Arc};
 
@@ -315,50 +316,6 @@ fn ellipsize(mut text: String, max_len: usize) -> String {
     }
     text.replace_range(prefix_len..text.len() - suffix_len, ellipsis);
     text
-}
-
-#[test]
-fn typing_whitespace_inside_a_function_should_not_invalidate_types() {
-    let (mut db, pos) = TestDB::with_position(
-        "
-        //- /lib.rs
-        fn foo() -> i32 {
-            $01 + 1
-        }
-    ",
-    );
-    {
-        let events = db.log_executed(|| {
-            let module = db.module_for_file(pos.file_id);
-            let crate_def_map = module.def_map(&db);
-            visit_module(&db, &crate_def_map, module.local_id, &mut |def| {
-                db.infer(def);
-            });
-        });
-        assert!(format!("{:?}", events).contains("infer"))
-    }
-
-    let new_text = "
-        fn foo() -> i32 {
-            1
-            +
-            1
-        }
-    "
-    .to_string();
-
-    db.set_file_text(pos.file_id, Arc::new(new_text));
-
-    {
-        let events = db.log_executed(|| {
-            let module = db.module_for_file(pos.file_id);
-            let crate_def_map = module.def_map(&db);
-            visit_module(&db, &crate_def_map, module.local_id, &mut |def| {
-                db.infer(def);
-            });
-        });
-        assert!(!format!("{:?}", events).contains("infer"), "{:#?}", events)
-    }
 }
 
 fn check_infer(ra_fixture: &str, expect: Expect) {
